requirements.txt
absl-py==2.1.0
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
python-dotenv
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
daal==2024.4.0
daal4py==2024.4.0
debugpy==1.8.0
boto3
decorator==5.1.1
Django==5.0.1
executing==2.0.0
filelock==3.15.4
waitress
flatbuffers==24.3.25
fonttools==4.43.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
idna==3.10
imageio==2.36.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.0
jupyter_client==8.4.0
jupyter_core==5.4.0
keras==3.5.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
mahotas==1.4.18
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
pymysql
namex==0.0.8
nest-asyncio==1.5.8
networkx==3.3
numpy
flask
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.12.1
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.25.5
psutil==5.9.6
pure-eval==0.2.2
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.7.0
pyzmq==25.1.1
requests==2.32.3
rich==13.9.1
scikit-image==0.24.0
scikit-learn==1.5.0
scikit-learn-intelex==2024.4.0
scipy
seaborn==0.13.2
setuptools==73.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.13.1
tbb==2021.12.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.9.20
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
wcwidth==0.2.8
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0